+++ src/models/lae.py
@@ -38,7 +38,7 @@ class LAE(BaseModel):
             W_np = np.linalg.solve(A_np, G_orig).astype(np.float32)
         except np.linalg.LinAlgError:
             print("[Warning] Singular matrix, using stronger regularization.")
-            A_np[np.diag_indices_from(A_np)] += 1e-4
+            A_np[np.diag_indices_from(A_np)] += np.float32(1e-4)
             W_np = np.linalg.solve(A_np, G_orig).astype(np.float32)
         
         self.weight_matrix = torch.tensor(W_np, dtype=torch.float32, device=self.device)
